requirements.txt
wandb==0.16.0
gymnasium==0.29.1
openai==0.28.0
hydra-core==1.3.2
omegaconf==2.3.0
poke-env==0.8.0
stable-baselines3==2.3.0